// emareach-frontend-next/types/api.ts
export interface Blog {
  id: string;
  title: string;
  slug: string;
  content?: string;
  excerpt?: string | null;
  author?: string | null;
  featured_image_url?: string | null;
  keywords?: string | null;
  tags?: string[] | null;
  status: string;
  published_at?: string | null;
  created_at: string;
  updated_at: string;
}